Don't delete image if not updating/deleting
git-svn-id: file:///home/svn/DVA2006@18669 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/dao/SetPartner.java
@@ -55,7 +55,7 @@ public class SetPartner extends DAO {
 					ps.setString(5, pi.getFormat().name().toLowerCase());
 					executeUpdate(ps, 1);
 				}
-			} else if (!isNew) {
+			} else if (!isNew && !pi.getHasImage()) {
 				try (PreparedStatement ps = prepare("DELETE FROM PARTNER_IMGS WHERE (ID=?)")) {
 					ps.setInt(1, pi.getID());
 					executeUpdate(ps, 0);
